backend/requirements.txt
Flask==2.2.5
flask-cors==3.0.10
pandas==2.1.2
numpy==1.24.3

pyyaml==6.0
requests==2.31.0
tqdm==4.66.1
tensorflow==2.13.0
scikit-learn==1.3.2
pmdarima
joblib==1.3.2
python-dateutil==2.8.2
pymongo==4.5.0
dnspython==2.4.2
